modularize image upload
